requirements.txt
fastapi==0.103.0
uvicorn==0.12.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==1.4.46
fastapi-mail==1.2.5
jinja2==3.1.2
pymysql==1.0.2
